move class to package for easy import
